Added basic calendar style
Added a border to see the calendar in light mode more clearly, and
centered and filled the screen.

TODO: Make more compatible for mobile and dark background in dark theme.
Could be difficult and limited by Open Web Calendar, but it is possible
to fetch from google calendar and use tailwindcss for a good look!